Signout Api & Admin Dashboard

+++ admin-app/src/actions/auth_action.js
@@ -59,9 +59,21 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
     return async dispatch =>{
-        localStorage.clear();
-        dispatch({
-            type: authConstants.LOGOUT_REQUEST
-        });
+
+        dispatch({type: authConstants.LOGOUT_REQUEST});
+
+        const res = await axios.post('/admin/sign-out');
+
+        if(res.status === 200){
+            localStorage.clear();
+            dispatch({ type: authConstants.LOGOUT_SUCCESS });
+        }else{
+            dispatch({
+                type: authConstants.LOGOUT_REQUEST,
+                payload: {
+                    error : res.data.error
+                }
+            });
+        }
     }
 }
